myproject/requirements.txt
asgiref==3.7.2
Django==3.2.10
django-environ==0.11.2
djangorestframework==3.13.1
Pillow==9.5.0
pytz==2024.2
sqlparse==0.4.4
typing-extensions==4.7.1
